Float problem solved

+++ Tramber/Assets/Resources/Scripts/LevelManager.cs
@@ -12,7 +12,7 @@ public class LevelManager : MonoBehaviour {
     public float absorbThreshouldRatio = 0.2f;
     public float absorbThreshouldTime = 0.4f;
 
-    public float dropMaxHealth = 100.0f;
+    public int dropMaxHealth = 100;
     public float timeToAbsorb = 3.0f;
